up endpoint catch notification from pg

+++ campaign/input.go
@@ -9,11 +9,11 @@ type GetCampaignDetailInput struct {
 
 // CreateCampaignInput is ...
 type CreateCampaignInput struct {
-	Name             string `json:"name" binding:"required"`
-	ShortDescription string `json:"short_description" binding:"required"`
-	Description      string `json:"description" binding:"required"`
-	GoalAmount       int    `json:"goal_amount" binding:"required"`
-	Perks            string `json:"perks" binding:"required"`
+	Name             string  `json:"name" binding:"required"`
+	ShortDescription string  `json:"short_description" binding:"required"`
+	Description      string  `json:"description" binding:"required"`
+	GoalAmount       float64 `json:"goal_amount" binding:"required"`
+	Perks            string  `json:"perks" binding:"required"`
 	User             user.User
 }
